Updating the Controllers and view resources  - Removed the CalendarController  - Removed the TimeSheetController  - Moved all views into the views/Default folder
