+++ src/AppBundle/Resources/views/Default/index.html.twig
@@ -1,4 +1,4 @@
-{% extends 'base.html.twig' %}
+{% extends "::base.html.twig" %}
 {% block title %}Howe - Status{% endblock %}
 {% block body %}
     <div class="row">
