--- src/AppBundle/Tests/Controller/CalendarControllerTest.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace AppBundle\Tests\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-
-class CalendarControllerTest extends WebTestCase
-{
-    public function testIndex()
-    {
-        $client = static::createClient();
-
-        $crawler = $client->request('GET', '/calendar');
-    }
-
-    public function testUpdate()
-    {
-        $client = static::createClient();
-
-        $crawler = $client->request('GET', '/utility/calendarUpdate');
-    }
-
-}
